+++ src/components/FinancialLedger/Setting.js
@@ -59,229 +59,6 @@ const SettingListTitle = (props) =>{
     );
 }
 
-//한달 예산 금액 설정
-const SetMonthlyBudget = (props) => {
-
-    const [isOpen, setIsOpen] = useState(false);
-
-    const openBudgetModalHandler = () => {
-        setIsOpen(true);
-    };
-    //취소버튼 클릭 시  값 초화
-    const closeBudgetModalHandler = () => {
-        setBudget(" ");
-        setIsOpen(false);
-    }
-
-    const [budget,setBudget] = useState(" ");
-
-    const onBudgetHandler = (event) => { 
-    setBudget(event.currentTarget.value);
-
-    }
-
-    const onBudgetSubmit = (event) => {
-        event.preventDefault();
-        let check = /^[0-9]+$/;
-        if(budget == " " && !check.test(budget) ){
-            document.getElementById('setBudget').innerHTML='<b>입력 형식이 올바르지 않습니다.<b>';
-            document.getElementById('setBudget').style.color='red';
-        }
-        else if(isNaN(budget) === false && budget > 0){
-            document.getElementById('showBudget').innerHTML=budget;
-            document.getElementById('showBudget').style.color='black';
-            setIsOpen(false)
-        }
-        else if(budget == 0){
-            document.getElementById('setBudget').innerHTML='<b>입력 값이 올바르지 않습니다.<b>';
-            document.getElementById('setBudget').style.color='red';
-        }
-        else{
-            document.getElementById('setBudget').innerHTML='<b>입력 형식이 올바르지 않습니다.<b>';
-            document.getElementById('setBudget').style.color='red';
-    }
-}
-
-    return(
-        <SettingListContentWrapper>
-            <span onClick={openBudgetModalHandler}>{props.content}</span>
-            {isOpen === false ? null : 
-            <ModalWrapper>
-                <StyledModal>
-                <h1>한달 예산 금액</h1>
-                <InputBox type="text" className="budget" value={budget} onChange={onBudgetHandler}></InputBox>
-                <ErrorMessageBox>
-                    <span id='setBudget'></span>
-                </ErrorMessageBox>
-                <ButtonBox>
-                    <button onClick={closeBudgetModalHandler} className="cancelButton">취소</button>
-                    <button onClick={onBudgetSubmit} className="checkButton">확인</button>
-                </ButtonBox>
-                </StyledModal>
-            </ModalWrapper>
-            }
-            <span id='showBudget'>--</span>
-        </SettingListContentWrapper>
-    );
-}
-
-//카테고리 설정 컴포넌트
-const SetCategory = (props) => {
-    const [isOpen, setIsOpen] = useState(false);
-
-    const [isClicked, setIsClicked] = useState(false);
-
-    const openCategoryModalHandler = () => {
-        setIsOpen(true)
-    };
-
-
-    //입력 박스 핸들러 조정
-    const openInputBoxHandler = () => {
-        setIsClicked(true)
-    }
-
-    //취소 버튼을 부르면 입력값도 모두 사라지도록 설정
-    const closeCategoryModalHandler = () => {
-        setIsOpen(false)
-    }
-
-    //취소를 클릭하면, 다시 아이콘이 뜨도록 설정하고 싶음
-     const closeCategoryInputHandler = () => {
-        setIsOpen(false)
-    }
-
-    const [category,SetCategory] = useState(" ");
-
-    //카레고리 입력값 설정
-    const onCategoryModalHandler = (event) => { 
-    SetCategory(event.currentTarget.value)
-    }
-
-    const onCategorySubmit = (event) => {
-        event.preventDefault();
-        //입력 받은 내용을 화면에 표시해주기
-            var category = document.getElementById("category");
-            var categoryContentDiv = document.getElementById("categoryContent");
-            var categoryValue = category.value; //입력 폼에 입력되는 카테고리 값
-
-            var categoryList = document.createElement("div");
-            categoryList.innerText = categoryValue;
-            categoryList.style.borderBottom = "1px solid black";
-            categoryList.style.width =  "300px";
-            categoryList.style.padding = "10px";
-            categoryList.style.textAlign = "left";
-            categoryList.style.fontSize = "15px";
-
-            categoryContentDiv.appendChild(categoryList);
-            
-    }
-
-
-
-    const StyledAddCircleOutlinedIcon = {
-        position: 'absolute',
-        bottom: '10px;',
-        right: '20px',
-        display : 'flex'
-    };
-    // + 아이콘, X 버튼, 삭제 아이콘 추가
-    // + 아이콘을 누르면 입력창이 뜨도록
-
-    return(
-        <SettingListContentWrapper>
-            {props.content}
-            <ChevronRightRoundedIcon onClick={openCategoryModalHandler}></ChevronRightRoundedIcon>
-            {isOpen === false ? null : 
-            <ModalWrapper>
-                <StyledModal>
-                <h1>{props.content}</h1>
-                <div className="Modal-close-btn" onClick={closeCategoryModalHandler}>&times;</div>
-                <AddCircleOutlinedIcon className="modalIcon" onclick={openInputBoxHandler} style={StyledAddCircleOutlinedIcon}></AddCircleOutlinedIcon>
-                        <InputBoxWrapper>
-                        <div id='categoryContent'>
-                                <DeleteIcon></DeleteIcon>
-                        </div>
-                             <input type="text" id="category" placeholder="카테고리를 입력하세요" onChange={onCategoryModalHandler}></input>
-                            <ButtonBox>
-                                <button onClick={closeCategoryInputHandler} className="cancelButton">취소</button>
-                                <button onClick={onCategorySubmit} className="checkButton">확인</button>
-                            </ButtonBox>
-                        </InputBoxWrapper>
-                </StyledModal>
-                </ModalWrapper>
-            }
-        </SettingListContentWrapper>
-    );
-}
-
-//지출 비율 설정 로직
-const SetExpenseRatio = (props) => {
-
-    const [isOpen, setIsOpen] = useState(false);
-
-    const openRatioModalHandler = () => {
-        setIsOpen(true);
-    };
-
-    //취소 버튼을 부르면 입력값도 모두 사라지도록 설정
-    const closeRatioModalHandler = () => {
-        setExpenseRatio(" ");
-        setIsOpen(false);
-    }
-
-    const [expenseRatio,setExpenseRatio] = useState(" ");
-
-    const onRatioModalHandler = (event) => { 
-        setExpenseRatio(event.currentTarget.value);
-    }
-
-    const onRatioSubmit = (event) => {
-        event.preventDefault();
-        let check = /^[0-9]+$/;
-        if(expenseRatio === null || !check.test(expenseRatio) ){
-            document.getElementById('setRatio').innerHTML='<b>입력 형식이 올바르지 않습니다.<b>';
-            document.getElementById('setRatio').style.color='red';
-        }
-        else if(isNaN(expenseRatio) === false){
-            document.getElementById('showRatio').innerHTML=expenseRatio;
-            document.getElementById('showRatio').style.color='black';
-            setIsOpen(false)
-        }
-        else if(expenseRatio < -1 || expenseRatio > 101){
-            document.getElementById('showRatio').innerHTML='<b>입력 값이 올바르지 않습니다.<b>';
-            document.getElementById('showRatio').style.color='black';
-        }
-        else{
-            document.getElementById('setRatio').innerHTML='<b>입력 형식이 올바르지 않습니다.<b>';
-            document.getElementById('setRatio').style.color='red';
-        }
-    }
-
-    return(
-        <SettingListContentWrapper>
-            <span onClick={openRatioModalHandler}>{props.content}</span>
-            {isOpen === false ? null : 
-            <ModalWrapper>
-                <StyledModal>
-                <h1>{props.content} 비율 설정</h1>
-                <InputBox type="text" className="ratio" value={expenseRatio} onChange={onRatioModalHandler}></InputBox><b>%</b>
-                <ErrorMessageBox>
-                    <span id='setRatio'></span>
-                </ErrorMessageBox>
-                <ButtonBox>
-                    <button onClick={closeRatioModalHandler} className="cancelButton">취소</button>
-                    <button onClick={onRatioSubmit} className="checkButton">확인</button>
-                </ButtonBox>
-                </StyledModal>
-            </ModalWrapper>
-            }
-            <span id='showRatio'>--</span>
-        </SettingListContentWrapper>
-    );
-}
-
-
     return(
         <div>
             <SettingList>
